+++ MORE/InternalInterfaces.cs
@@ -6,5 +6,5 @@ namespace Shockah.MORE;
 internal interface IRegisterable
 {
 	static abstract void Register(IPluginPackage<IModManifest> package, IModHelper helper);
-	virtual static void UpdateSettings(IPluginPackage<IModManifest> package, IModHelper helper, ProfileSettings settings) { }
+	static virtual void UpdateSettings(IPluginPackage<IModManifest> package, IModHelper helper, ProfileSettings settings) { }
 }